fix : devide && conquer
