+++ schema.js
@@ -4,7 +4,7 @@ import { mergeResolvers, mergeTypeDefs } from "@graphql-tools/merge";
 
 const loadedTypes = loadFilesSync(`${__dirname}/**/*.typeDefs.js`);
 const loadedResolvers = loadFilesSync(
-  `${__dirname}/**/*.{queries,mutations}.js`
+  `${__dirname}/**/*.{queries,resolvers}.js`
 );
 
 const typeDefs = mergeTypeDefs(loadedTypes);
