+++ users/createAccount/createAccount.resolvers.js
@@ -1,0 +1,41 @@
+import bcrypt from "bcrypt";
+import client from "../../client";
+
+export default {
+  Mutation: {
+    createAccount: async (
+      _,
+      { firstName, lastName, username, email, password }
+    ) => {
+      try {
+        const existingUser = await client.user.findFirst({
+          where: {
+            OR: [
+              {
+                username,
+              },
+              {
+                email,
+              },
+            ],
+          },
+        });
+        if (existingUser) {
+          throw new Error("This username/password is already taken.");
+        }
+        const uglyPassword = await bcrypt.hash(password, 10);
+        return client.user.create({
+          data: {
+            username,
+            email,
+            firstName,
+            lastName,
+            password: uglyPassword,
+          },
+        });
+      } catch (e) {
+        return e;
+      }
+    },
+  },
+};
