+++ users/createAccount/createAccount.typeDefs.js
@@ -1,0 +1,17 @@
+import { gql } from "apollo-server";
+
+export default gql`
+  type CreateAccountResult {
+    ok: Boolean!
+    error: String
+  }
+  type Mutation {
+    createAccount(
+      firstName: String!
+      lastName: String
+      username: String!
+      email: String!
+      password: String!
+    ): CreateAccountResult!
+  }
+`;
