+++ users/editProfile/editProfile.resolvers.js
@@ -1,0 +1,7 @@
+export default {
+  Mutation: {
+    editProfile: () => {
+      console.log("hi");
+    },
+  },
+};
