+++ users/editProfile/editProfile.typeDefs.js
@@ -1,0 +1,17 @@
+import { gql } from "apollo-server";
+
+export default gql`
+  type EditProfileResult {
+    ok: Boolean!
+    error: String
+  }
+  type Mutation {
+    editProfile(
+      firstName: String
+      lastName: String
+      username: String
+      email: String
+      password: String
+    ): EditProfileResult!
+  }
+`;
