+++ users/login/login.resolvers.js
@@ -1,0 +1,29 @@
+import bcrypt from "bcrypt";
+import jwt from "jsonwebtoken";
+import client from "../../client";
+
+export default {
+  Mutation: {
+    login: async (_, { username, password }) => {
+      const user = await client.user.findFirst({ where: { username } });
+      if (!user) {
+        return {
+          ok: false,
+          error: "User not found.",
+        };
+      }
+      const passwordOk = await bcrypt.compare(password, user.password);
+      if (!passwordOk) {
+        return {
+          ok: false,
+          error: "Incorrect password.",
+        };
+      }
+      const token = await jwt.sign({ id: user.id }, process.env.SECRET_KEY);
+      return {
+        ok: true,
+        token,
+      };
+    },
+  },
+};
