+++ users/login/login.typeDefs.js
@@ -1,0 +1,12 @@
+import { gql } from "apollo-server";
+
+export default gql`
+  type LoginResult {
+    ok: Boolean!
+    token: String
+    error: String
+  }
+  type Mutation {
+    login(username: String!, password: String!): LoginResult!
+  }
+`;
