+++ users/seeProfile/seeProfile.typeDefs.js
@@ -1,0 +1,7 @@
+import { gql } from "apollo-server";
+
+export default gql`
+  type Query {
+    seeProfile(username: String!): User
+  }
+`;
