+++ users/users.typeDefs.js
@@ -10,22 +10,4 @@ export default gql`
     createdAt: String!
     updatedAt: String!
   }
-  type LoginResult {
-    ok: Boolean!
-    token: String
-    error: String
-  }
-  type Mutation {
-    createAccount(
-      firstName: String!
-      lastName: String
-      username: String!
-      email: String!
-      password: String!
-    ): User
-    login(username: String!, password: String!): LoginResult!
-  }
-  type Query {
-    seeProfile(username: String!): User
-  }
 `;
